+++ CHANGELOG.md
@@ -5,10 +5,11 @@
 ### Added
 
 - added parameter `use_tls`. It allows to activate or not the tls use within smtp connection. Default value is `True` for backward compatibility.
+- added the check to ensure the SMTP server supports the auth extension before login
 
 ### Changed
 
-- nothing changed
+- change function `get_connection` badly named `sender_email` parameter
 
 ### Removed